requirements.txt
discord.py==2.4.1
python-dotenv
yt-dlp
PyNaCl
certifi==2024.2.2